link: ensure fan on at cold start, off in suspend
The previous commit attempted to ensure the fan was on
at cold start by enabling it any time the RPM was set,
but resulted in the fan remaining on when suspended,
pulsing on and off in a continuous cycle.

Fix this by instead enabling the fan and setting the RPM
mode to auto in init function and revert previous enabling
within the set rpm function

Test: verify fan enabled on cold reset and off on suspend/
shutdown via serial console.

Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>